Add support for external file loading as modules
Adding modules is done via menu.lst in the grub configuration, while the
new modules should be in isodir/modules. Currently only supported feature
is printing modules' content to screen in ascii.
Future features planned are executing a module with machine instructions
in kernel-space. This is very insecure, and the external program will
have to ret on it's own, but this is only for testing and PoC.